requirements.txt
# prajekt Backend Dependencies
# Root reference file for deployment systems

# Framework
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-SocketIO==5.3.5
Flask-JWT-Extended==4.5.3
Flask-Migrate==4.0.5
Flask-Cors==4.0.0
Werkzeug==3.0.1

# Database
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.0

# WebSocket & Real-time
python-socketio==5.10.0
python-engineio==4.8.0

# AI / ML (latest stable version)
google-generativeai>=0.8.0

# Environment
python-dotenv==1.0.0

# Production Server
gunicorn==21.2.0
eventlet==0.33.3

# Utilities
requests==2.31.0
